const validation: remove check for mutable refs in final value of const
